requirements.txt
numpy
scikit-learn
pycm
nltk
aspell-python-py3
tweet-preprocessor